Revert "scaling"
This reverts commit cd212b2214f33f49f5e5dbe31018385c83a2df00.

+++ Core/include/Acts/Material/MaterialSlab.hpp
@@ -59,7 +59,7 @@ class MaterialSlab {
   MaterialSlab& operator=(const MaterialSlab&) = default;
 
   /// Scale the material thickness by the given factor.
-  void scaleThickness(double scale);
+  void scaleThickness(float scale);
 
   /// Check if the material is valid, i.e. it is finite and not vacuum.
   bool isValid() const { return m_material.isValid() && (0.0f < m_thickness); }
