+++ Core/src/Material/VolumeMaterialMapper.cpp
@@ -253,7 +253,7 @@ void Acts::VolumeMaterialMapper::createExtraHits(
   // Computing the extra hits properties based on the mappingStep length
   int volumeStep =
       static_cast<int>(std::floor(properties.thickness() / m_cfg.mappingStep));
-  double remainder = properties.thickness() - m_cfg.mappingStep * volumeStep;
+  float remainder = properties.thickness() - m_cfg.mappingStep * volumeStep;
   properties.scaleThickness(m_cfg.mappingStep / properties.thickness());
   direction = direction * (m_cfg.mappingStep / direction.norm());
